+++ MoodWcfService/IServices/IEvaluationService.cs
@@ -7,7 +7,7 @@ using System.ServiceModel.Web;
 using System.Text;
 using MoodWcfService.Entities;
 
-namespace MoodWcfService.Services
+namespace MoodWcfService.IServices
 {
     // NOTE: You can use the "Rename" command on the "Refactor" menu to change the interface name "IEvaluationService" in both code and config file together.
     [ServiceContract]
