+++ MoodWcfService/IServices/IRestaurantService.cs
@@ -7,43 +7,44 @@ using System.ServiceModel.Web;
 using System.Text;
 using MoodWcfService.Entities;
 
-namespace MoodWcfService.Services
+namespace MoodWcfService.IServices
 {
-    // NOTE: You can use the "Rename" command on the "Refactor" menu to change the interface name "IUserService" in both code and config file together.
+    // NOTE: You can use the "Rename" command on the "Refactor" menu to change the interface name "IRestaurantService" in both code and config file together.
     [ServiceContract]
-    public interface IUserService
+    public interface IRestaurantService
     {
         [OperationContract]
         [WebInvoke(Method = "GET",
             ResponseFormat = WebMessageFormat.Json,
-            UriTemplate = "/GetAllUsers")]
-        List<User_> GetAllUsers();
+            UriTemplate = "/GetAllRestaurants")]
+        List<Restaurant_> GetAllRestaurants();
 
         [OperationContract]
         [WebInvoke(Method = "GET",
             ResponseFormat = WebMessageFormat.Json,
-            UriTemplate = "/GetUserById/{userId}")]
-        User_ GetUserById(string userId);
+            UriTemplate = "/GetRestaurantById/{restId}")]
+        Restaurant_ GetRestaurantById(string restId);
 
         [OperationContract]
         [WebInvoke(Method = "POST",
             RequestFormat = WebMessageFormat.Json,
             ResponseFormat = WebMessageFormat.Json,
             UriTemplate = "/Create")]
-        bool CreateUser(User_ user);
+        bool CreateRestaurant(Restaurant_ restaurant);
 
         [OperationContract]
         [WebInvoke(Method = "PUT",
             RequestFormat = WebMessageFormat.Json,
             ResponseFormat = WebMessageFormat.Json,
             UriTemplate = "/Edit")]
-        bool EditUser(User_ user);
+        bool EditRestaurant(Restaurant_ restaurant);
 
         [OperationContract]
         [WebInvoke(Method = "DELETE",
             RequestFormat = WebMessageFormat.Json,
             ResponseFormat = WebMessageFormat.Json,
             UriTemplate = "/Delete")]
-        bool DeleteUser(User_ user);
+        bool DeleteRestaurant(Restaurant_ restaurant);
     }
 }
+
