+++ MoodWcfService/Web.config
@@ -22,7 +22,7 @@
                behaviorConfiguration="MoodWCFService_Behavior" >
         <endpoint address=""
                   binding="webHttpBinding"
-                  contract="MoodWcfService.Services.IUserService" >
+                  contract="MoodWcfService.IServices.IUserService" >
         </endpoint>
       </service>
 
@@ -30,7 +30,7 @@
                behaviorConfiguration="MoodWCFService_Behavior" >
         <endpoint address=""
                   binding="webHttpBinding"
-                  contract="MoodWcfService.Services.IRestaurantService" >
+                  contract="MoodWcfService.IServices.IRestaurantService" >
         </endpoint>
       </service>
 
@@ -38,7 +38,7 @@
                behaviorConfiguration="MoodWCFService_Behavior" >
         <endpoint address=""
                   binding="webHttpBinding"
-                  contract="MoodWcfService.Services.IEvaluationService" >
+                  contract="MoodWcfService.IServices.IEvaluationService" >
         </endpoint>
       </service>
